proposed system
The proposed system of project is that provides easy way of reserving a parking space online using web portal. It overcomes the problem of finding a parking space in areas that unnecessary consumes time. Hence, this project offers a web application based reservation system where users can view various parking spaces and select nearby or specific area of their choice to view whether space is available or not. If the booking space is available, then user can book it for specific time slot. The booked space will be marked and will not be available for anyone else for the specified time.
Advantages:
It is representing clearly with maps.
It has authorized parking slots with authorized address.
It is so easy to use with simple UI.
It has reserved for particular timing.