existing project
We presently model the internet parking spot sharing issue as a two-sided market, where the PSPs go about as venders and the PSCs go about as
purchasers. The market will be set off when there are purchasers and dealers all the while. The framework works in a period opened style. The
time allotment is set by the representative, and in this article, we consider the schedule opening is 60 minutes. Likewise, the PSPs and the PSCs can
show up and withdraw from the market progressively, without advance information on the bartering.
Disadvantages:
It is not represent parking spaces and slots.
It is does not have authorized parking slots.
It requires large database.
It does not have reserved for particular timing.